Add starting removed locations to the db
